Commit 1 of ExtendedHelper
Created HuskyHelper which extends Helper and takes care of remembering the coder and the postSorter; also keeps track of the size (N) and the long array.
WARNING: this version is about 3% slower than the master branch version so it should NOT be merged until performance is resolved.
Also, the code needs significant cleanup before being good code.